Move registry cleaner to helpers
